Added several functions for better manipulation
Added several ways to configure and bind library. Now you can bind for : Window, WindowTab, GUIManager, TextManager via following:
PINGUI::window - for window functions 
PINGUI::windowTab - for window tab functions
PINGUI::GUI - accesing the GUI management functions
PINGUI::TEXT - accessing the text functions

Windows are now created using the new structure called windowDef, which stores all needed informations. Added the possibility to add a name tags for different windows so you dont have to keep the pointer to them if you wanted to use them in later stage, but simply binding them via their name tag.

Also fixed the bug where it was possible to collide with the main gui manager even through the windows.  Updated the demo cpp file so it supports the updated syntax.

+++ PinGUI/GL_Vertex.h
@@ -59,7 +59,7 @@ namespace PinGUI{
             float x;
             float y;
             VertexPosition();
-            VertexPosition(int X, int Y):
+            VertexPosition(float X, float Y):
                 x(X),
                 y(Y){
                 };
